refactor(swoole): Move AtomicCounter to Memory\Atomic namespace
